Prequantized feature variables are now Doubles

+++ src/system/method/feature/Variable.java
@@ -5,18 +5,18 @@ import settings.settingsvariables.SettingsMethodVariable;
 public class Variable {
  	
 	public SettingsMethodVariable variableSettings;
-	protected Long prequantizedValue; // TODO Jesse - make these Double
+	protected Double prequantizedValue; // TODO Jesse - make these Double
 //	protected Long quantizedValue;
 	
 	public Variable(SettingsMethodVariable variableSettings){
 		this.variableSettings = variableSettings;
 	}
 
-	public Long getPrequantizedValue() {
+	public Double getPrequantizedValue() {
 		return prequantizedValue;
 	}
 
-	public void setPrequantizedValue(Long prequantizedValue) {
+	public void setPrequantizedValue(Double prequantizedValue) {
 		this.prequantizedValue = prequantizedValue;
 //		this.setQuantizedValue(variableSettings.findBin(prequantizedValue));
 	}
